+++ src/Oro/Bundle/FrontendBundle/Resources/public/js/app/datafilter/frontend-multiselect-decorator.js
@@ -187,7 +187,7 @@ define(function(require, exports, module) {
         addAdditionalClassesForContainer: function(widget) {
             if (this.parameters.additionalClass) {
                 widget
-                    .removeAttr('class')
+                    .attr('class', null)
                     .addClass(`${this.parameters.themeName} dropdown-menu`);
             }
         },
@@ -198,7 +198,7 @@ define(function(require, exports, module) {
          */
         setDropdownHeaderDesign: function(instance) {
             instance.header
-                .removeAttr('class')
+                .attr('class', null)
                 .addClass('datagrid-manager__header');
         },
 
@@ -217,7 +217,7 @@ define(function(require, exports, module) {
             instance.header
                 .find('.ui-multiselect-filter')
                 .contents()
-                .eq(0)
+                .first()
                 .filter((i, el) => el.nodeType === Node.TEXT_NODE)
                 .wrap(
                     $('<h5></h5>', {'class': 'datagrid-manager__title'})
@@ -225,7 +225,7 @@ define(function(require, exports, module) {
 
             instance.header
                 .find('.ui-multiselect-filter')
-                .removeAttr('class');
+                .attr('class', null);
 
             instance.header
                 .find('.ui-multiselect-close')